All data passed around in Horde internally should be UTF-8.

+++ framework/Ldap/test/Horde/Ldap/SearchTest.php
@@ -13,6 +13,9 @@ class Horde_Ldap_SearchTest extends Horde_Ldap_TestBase
 {
     public static function tearDownAfterClass()
     {
+        if (!self::$ldapcfg) {
+            return;
+        }
         $ldap = new Horde_Ldap(self::$ldapcfg['server']);
         $ldap->delete('ou=Horde_Ldap_Test_search1,' . self::$ldapcfg['server']['basedn']);
         $ldap->delete('ou=Horde_Ldap_Test_search2,' . self::$ldapcfg['server']['basedn']);
